refactor: moved types and added unit tests
