+++ src/components/Popover/Popover.constants.ts
@@ -1,6 +1,33 @@
-import type { Round } from './Popover.types';
+import type { Round, Color, PlacementType } from './Popover.types';
 
 const CLASS_PREFIX = 'md-popover';
+
+const COLORS: Record<string, Color> = {
+  PRIMARY: 'primary',
+  SECONDARY: 'secondary',
+  TERTIARY: 'tertiary',
+  QUATERNARY: 'quaternary',
+};
+
+const PLACEMENTS: Record<string, PlacementType> = {
+  TOP: 'top',
+  TOP_START: 'top-start',
+  TOP_END: 'top-end',
+  RIGHT: 'right',
+  RIGHT_START: 'right-start',
+  RIGHT_END: 'right-end',
+  BOTTOM: 'bottom',
+  BOTTOM_START: 'bottom-start',
+  BOTTOM_END: 'bottom-end',
+  LEFT: 'left',
+  LEFT_START: 'left-start',
+  LEFT_END: 'left-end',
+
+  // choose the side with most space:
+  AUTO: 'auto',
+  AUTO_START: 'auto-start',
+  AUTO_END: 'auto-end',
+};
 
 const ROUNDS: Record<number, Round> = {
   0: 0,
@@ -23,4 +50,4 @@ const STYLE = {
   arrowWrapper: `${CLASS_PREFIX}-arrow-wrapper`,
 };
 
-export { CLASS_PREFIX, DEFAULTS, STYLE };
+export { CLASS_PREFIX, DEFAULTS, STYLE, COLORS, PLACEMENTS };
